+++ src/apps/refactor_proto/refactor_proto/CMakeLists.txt
@@ -17,6 +17,7 @@ set(${TARGET_NAME}_HEADERS
     Logging.h
     Material.h
     Model.h
+    Pass.h
     Profiler.h
     Profiling.h
     RenderGraph.h
@@ -25,6 +26,7 @@ set(${TARGET_NAME}_HEADERS
     Scene.h
     SceneGui.h
     SetupDrawing.h
+    TheGraph.h
     Time.h
 
     providers/ProviderApi.h
@@ -40,12 +42,14 @@ set(${TARGET_NAME}_SOURCES
     IntrospectProgram.cpp
     Loader.cpp
     Logging.cpp
+    Pass.cpp
     Profiler.cpp
     Profiling.cpp
     RenderGraph.cpp
     Scene.cpp
     SceneGui.cpp
     SetupDrawing.cpp
+    TheGraph.cpp
 
     providers/ProviderCpu.cpp
     providers/ProviderGL.cpp
