+++ src/apps/refactor_proto/refactor_proto/RenderGraph.cpp
@@ -1,13 +1,15 @@
 #include "RenderGraph.h"
 
+// TODO Ad 2023/10/18: Should get rid of this repeated implementation
+#include "RendererReimplement.h" 
+
 #include "Cube.h"
 #include "GlApi.h"
 #include "Json.h"
 #include "Loader.h"
 #include "Logging.h"
+#include "Pass.h"
 #include "Profiling.h"
-#include "RendererReimplement.h" // TODO Ad 2023/10/18: Should get rid of this repeated implementation
-#include "SetupDrawing.h"
 
 #include <handy/vector_utils.h>
 
@@ -96,8 +98,6 @@ constexpr float gMinFarZ{-25.f};
 [[maybe_unused]] constexpr float gRadialDistancingFactor{1.5f};
 /// @brief The far plane will be at least at this factor times the model depth.
 constexpr float gDepthFactor{20.f};
-
-constexpr std::size_t gMaxDrawInstances = 2048;
 
 namespace {
 
@@ -211,79 +211,6 @@ namespace {
     }
 
 
-    void draw(const PassCache & aPassCache,
-              const RepositoryUbo & aUboRepository,
-              const RepositoryTexture & aTextureRepository)
-    {
-        //TODO Ad 2023/08/01: META todo, should we have "compiled state objects" (a-la VAO) for interface bocks, textures, etc
-        // where we actually map a state setup (e.g. which texture name to which image unit and which sampler)
-        // those could be "bound" before draw (potentially doing some binds and uniform setting, but not iterating the program)
-        // (We could even separate actual texture from the "format", allowing to change an underlying texture without revisiting the program)
-        // This would address the warnings repetitions (only issued when the compiled state is (re-)generated), and be better for perfs.
-
-        GLuint firstInstance = 0; 
-        for (const DrawCall & aCall : aPassCache.mCalls)
-        {
-            PROFILER_SCOPE_RECURRING_SECTION("drawcall_iteration", CpuTime);
-
-            PROFILER_PUSH_RECURRING_SECTION("discard_2", CpuTime);
-            const IntrospectProgram & selectedProgram = *aCall.mProgram;
-            const graphics::VertexArrayObject & vao = *aCall.mVao;
-            PROFILER_POP_RECURRING_SECTION;
-
-            // TODO Ad 2023/10/05: #perf #azdo 
-            // Only change what is necessary, instead of rebiding everything each time.
-            // Since we sorted our draw calls, it is very likely that program remain the same, and VAO changes.
-            {
-                PROFILER_PUSH_RECURRING_SECTION("bind_VAO", CpuTime);
-                graphics::ScopedBind vaoScope{vao};
-                PROFILER_POP_RECURRING_SECTION;
-                
-                {
-                    PROFILER_SCOPE_RECURRING_SECTION("set_buffer_backed_blocks", CpuTime);
-                    // TODO #repos This should be consolidated
-                    RepositoryUbo uboRepo{aUboRepository};
-                    if(aCall.mCallContext)
-                    {
-                        RepositoryUbo callRepo{aCall.mCallContext->mUboRepo};
-                        uboRepo.merge(callRepo);
-                    }
-                    setBufferBackedBlocks(selectedProgram, uboRepo);
-                }
-
-                {
-                    PROFILER_SCOPE_RECURRING_SECTION("set_textures", CpuTime);
-                    // TODO #repos This should be consolidated
-                    RepositoryTexture textureRepo{aTextureRepository};
-                    if(aCall.mCallContext)
-                    {
-                        RepositoryTexture callRepo{aCall.mCallContext->mTextureRepo};
-                        textureRepo.merge(callRepo);
-                    }
-                    setTextures(selectedProgram, textureRepo);
-                }
-
-                PROFILER_PUSH_RECURRING_SECTION("bind_program", CpuTime);
-                graphics::ScopedBind programScope{selectedProgram};
-                PROFILER_POP_RECURRING_SECTION;
-
-                {
-                    // TODO Ad 2023/08/23: Measuring GPU time here has a x2 impact on cpu performance
-                    // Can we have efficient GPU measures?
-                    PROFILER_SCOPE_RECURRING_SECTION("glDraw_call", CpuTime/*, GpuTime*/);
-                    
-                    gl.MultiDrawElementsIndirect(
-                        aCall.mPrimitiveMode,
-                        aCall.mIndicesType,
-                        (void *)(firstInstance * sizeof(DrawElementsIndirectCommand)),
-                        aCall.mPartCount,
-                        sizeof(DrawElementsIndirectCommand));
-                }
-            }
-            firstInstance += aCall.mPartCount;
-        }
-    }
-
     void populatePartList(PartList & aPartList, const Node & aNode, const Pose & aParentPose, const Material * aMaterialOverride)
     {
         const Instance & instance = aNode.mInstance;
@@ -321,291 +248,6 @@ namespace {
     }
 
 
-    /// @brief Find the program satisfying all annotations.
-    /// @return The first program in aEffect for which all annotations are matching (i.e. present and same value),
-    /// or a null handle otherwise.
-    template <class T_iterator>
-    Handle<ConfiguredProgram> getProgram(const Effect & aEffect,
-                                         T_iterator aAnnotationsBegin, T_iterator aAnnotationsEnd)
-    {
-        for (const Technique & technique : aEffect.mTechniques)
-        {
-            if(std::all_of(aAnnotationsBegin, aAnnotationsEnd,
-                    [&technique](const Technique::Annotation & aRequiredAnnotation)
-                    {
-                        const auto & availableAnnotations = technique.mAnnotations;
-                        if(auto found = availableAnnotations.find(aRequiredAnnotation.mCategory);
-                            found != availableAnnotations.end())
-                        {
-                            return (found->second == aRequiredAnnotation.mValue);
-                        }
-                        return false;
-                    }))
-            {
-                return technique.mConfiguredProgram;
-            }
-        }
-
-        return nullptr;
-    }
-
-
-    /// @brief Specializes getProgram, returning first program matching provided pass name.
-    Handle<ConfiguredProgram> getProgramForPass(const Effect & aEffect, StringKey aPassName)
-    {
-        const std::array<Technique::Annotation, 2> annotations{/*aggregate init of std::array*/{/*aggregate init of inner C-array*/
-            {"pass", aPassName},
-            {"pass", aPassName},
-        }};
-        return getProgram(aEffect, annotations.begin(), annotations.end());
-    }
-    
-
-    /// @brief Returns the VAO matching the given program and part.
-    /// Under the hood, the VAO is cached.
-    Handle<graphics::VertexArrayObject> getVao(const ConfiguredProgram & aProgram,
-                                               const Part & aPart,
-                                               Storage & aStorage)
-    {
-        // This is a common mistake, it would be nice to find a safer way
-        assert(aProgram.mConfig);
-
-        // Note: the config is via "handle", hosted by in cache that is mutable, so loosing the constness is correct.
-        return
-            [&, &entries = aProgram.mConfig->mEntries]() 
-            -> graphics::VertexArrayObject *
-            {
-                if(auto foundConfig = std::find_if(entries.begin(), entries.end(),
-                                                [&aPart](const auto & aEntry)
-                                                {
-                                                    return aEntry.mVertexStream == aPart.mVertexStream;
-                                                });
-                    foundConfig != entries.end())
-                {
-                    return foundConfig->mVao;
-                }
-                else
-                {
-                    aStorage.mVaos.push_back(prepareVAO(aProgram.mProgram, *aPart.mVertexStream));
-                    entries.push_back(
-                        ProgramConfig::Entry{
-                            .mVertexStream = aPart.mVertexStream,
-                            .mVao = &aStorage.mVaos.back(),
-                        });
-                    return entries.back().mVao;
-                }
-            }();
-    }
-
-
-    // TODO Ad 2023/10/05: for OpenGL resource, maybe we should directly use the GL name?
-    // or do we want to reimplement Handle<> in term of index in storage containers?
-    /// @brief Hackish class, associating a resource to an integer value.
-    ///
-    /// The value is suitable to compose a DrawEntry::Key via bitwise operations.
-    template <class T_resource, unsigned int N_valueBits>
-    struct ResourceIdMap
-    {
-        static_assert(N_valueBits <= 16);
-        static constexpr std::uint16_t gMaxValue = (std::uint16_t)((1 << N_valueBits) - 1);
-
-        ResourceIdMap()
-        {
-            // By adding the nullptr as the zero value, this helps debugging
-            get(nullptr);
-            assert(get(nullptr)== 0);
-        }
-
-        /// @brief Get the value associated to provided resource.
-        /// A given map instance will always return the same value for the same resource.
-        std::uint16_t get(Handle<T_resource> aResource)
-        {
-            assert(mResourceToId.size() < gMaxValue);
-            auto [iterator, didInsert] = mResourceToId.try_emplace(aResource,
-                                                                   (std::uint16_t)mResourceToId.size());
-            return iterator->second;
-        }
-
-        /// @brief Return the resource from it value.
-        Handle<T_resource> reverseLookup(std::uint16_t aValue)
-        {
-            if(auto found = std::find_if(mResourceToId.begin(), mResourceToId.end(),
-                                         [aValue](const auto & aPair)
-                                         {
-                                             return aPair.second == aValue;
-                                         });
-               found != mResourceToId.end())
-            {
-                return found->first;
-            }
-            throw std::logic_error{"The provided value is not present in this ResourceIdMap."};
-        }
-
-        std::unordered_map<Handle<T_resource>, std::uint16_t> mResourceToId;
-    };
-
-
-    /// @brief Associate an integer key to a part.
-    /// Used to sort an array with one entry for each part, by manually composing a key by sort dimensions.
-    /// @see https://realtimecollisiondetection.net/blog/?p=86
-    struct DrawEntry
-    {
-        using Key = std::uint64_t;
-        static constexpr Key gInvalidKey = std::numeric_limits<Key>::max();
-
-        /// @brief Order purely by key.
-        /// @param aRhs 
-        /// @return 
-        bool operator<(const DrawEntry & aRhs) const
-        {
-            return mKey < aRhs.mKey;
-        }
-
-        Key mKey = 0;
-        std::size_t mPartListIdx;
-    };
-
-
-    constexpr std::uint64_t makeMask(unsigned int aBitCount)
-    {
-        assert(aBitCount <= 64);
-        return (1 << aBitCount) - 1;
-    }
-
-
-    PassCache preparePass(StringKey aPass,
-                          const PartList & aPartList,
-                          Storage & aStorage)
-    {
-        PROFILER_SCOPE_RECURRING_SECTION("prepare_pass", CpuTime);
-
-        constexpr unsigned int gProgramIdBits = 16;
-        constexpr std::uint64_t gProgramIdMask = makeMask(gProgramIdBits);
-        using ProgramId = std::uint16_t;
-        assert(aStorage.mPrograms.size() < (1 << gProgramIdBits));
-
-        //constexpr unsigned int gPrimitiveModes = 12; // I counted 12 primitive modes
-        //unsigned int gPrimitiveModeIdBits = (unsigned int)std::ceil(std::log2(gPrimitiveModes));
-
-        constexpr unsigned int gVaoBits = 10;
-        constexpr std::uint64_t gVaoIdMask = makeMask(gVaoBits);
-        assert(aStorage.mVaos.size() < (1 << gVaoBits));
-        using VaoId = std::uint16_t;
-
-        constexpr unsigned int gMaterialContextBits = 10;
-        constexpr std::uint64_t gMaterialContextIdMask = makeMask(gMaterialContextBits);
-        assert(aStorage.mMaterialContexts.size() < (1 << gMaterialContextBits));
-        using MaterialContextId = std::uint16_t;
-
-        ResourceIdMap<ConfiguredProgram, gProgramIdBits> programToId;
-        ResourceIdMap<graphics::VertexArrayObject, gVaoBits> vaoToId;
-        ResourceIdMap<MaterialContext, gMaterialContextBits> materialContextToId;
-
-        //
-        // For each part, associated it to its sort key and store them in an array
-        // (this will already prune parts for which there is no program for aPass)
-        //
-        std::vector<DrawEntry> entries;
-        entries.reserve(aPartList.mParts.size());
-
-        for (std::size_t partIdx = 0; partIdx != aPartList.mParts.size(); ++partIdx)
-        {
-            const Part & part = *aPartList.mParts[partIdx];
-            const Material & material = *aPartList.mMaterials[partIdx];
-
-            // TODO Ad 2023/10/05: Also add those as sorting dimensions (to the LSB)
-            assert(part.mPrimitiveMode == GL_TRIANGLES);
-            assert(part.mVertexStream->mIndicesType == GL_UNSIGNED_INT);
-            
-            if(Handle<ConfiguredProgram> configuredProgram = getProgramForPass(*material.mEffect, aPass);
-               configuredProgram)
-            {
-                Handle<graphics::VertexArrayObject> vao = getVao(*configuredProgram, part, aStorage);
-
-                ProgramId programId = programToId.get(configuredProgram);
-                VaoId vaoId = vaoToId.get(vao);
-                MaterialContextId materialContextId = materialContextToId.get(material.mContext);
-
-                std::uint64_t key = vaoId ;
-                key |= materialContextId << gVaoBits;
-                key |= programId << (gVaoBits + gMaterialContextBits);
-
-                entries.push_back(DrawEntry{.mKey = key, .mPartListIdx = partIdx});
-            }
-        }
-
-        //
-        // Sort the entries
-        //
-        {
-            PROFILER_SCOPE_RECURRING_SECTION("sort_draw_entries", CpuTime);
-            std::sort(entries.begin(), entries.end());
-        }
-
-        //
-        // Traverse the sorted array to generate the actual draw commands and draw calls
-        //
-        PassCache result;
-        DrawEntry::Key drawKey = DrawEntry::gInvalidKey;
-        for(const DrawEntry & entry : entries)
-        {
-            const Part & part = *aPartList.mParts[entry.mPartListIdx];
-            const VertexStream & vertexStream = *part.mVertexStream;
-
-            // If true: this is the start of a new DrawCall
-            if(entry.mKey != drawKey)
-            {
-                // Record the new drawkey
-                drawKey = entry.mKey;
-
-                // Push the new DrawCall
-                result.mCalls.push_back(
-                    DrawCall{
-                        .mPrimitiveMode = part.mPrimitiveMode,
-                        .mIndicesType = vertexStream.mIndicesType,
-                        .mProgram = 
-                            &programToId.reverseLookup(gProgramIdMask & (drawKey >> (gVaoBits + gMaterialContextBits)))
-                                ->mProgram,
-                        .mVao = vaoToId.reverseLookup(gVaoIdMask & drawKey),
-                        .mCallContext = materialContextToId.reverseLookup(gMaterialContextIdMask & (drawKey >> gVaoBits)),
-                        .mPartCount = 0,
-                    }
-                );
-            }
-            
-            // Increment the part count of current DrawCall
-            ++result.mCalls.back().mPartCount;
-
-            const std::size_t indiceSize = graphics::getByteSize(vertexStream.mIndicesType);
-
-            // Indirect draw commands do not accept a (void*) offset, but a number of indices (firstIndex).
-            // This means the offset into the buffer must be aligned with the index size.
-            assert((vertexStream.mIndexBufferView.mOffset %  indiceSize) == 0);
-
-            result.mDrawCommands.push_back(
-                DrawElementsIndirectCommand{
-                    .mCount = part.mIndicesCount,
-                    // TODO Ad 2023/09/26: #bench Is it worth it to group identical parts and do "instanced" drawing?
-                    // For the moment, draw a single instance for each part (i.e. no instancing)
-                    .mInstanceCount = 1,
-                    .mFirstIndex = (GLuint)(vertexStream.mIndexBufferView.mOffset / indiceSize)
-                                    + part.mIndexFirst,
-                    .mBaseVertex = part.mVertexFirst,
-                    .mBaseInstance = (GLuint)result.mDrawInstances.size(), // pushed below
-                }
-            );
-
-            const Material & material = *aPartList.mMaterials[entry.mPartListIdx];
-
-            result.mDrawInstances.push_back(DrawInstance{
-                .mInstanceTransformIdx = aPartList.mTransformIdx[entry.mPartListIdx],
-                .mMaterialIdx = (GLsizei)material.mPhongMaterialIdx,
-            });
-        }
-
-        return result;
-    }
-
 } // unnamed namespace
 
 
@@ -617,22 +259,6 @@ PartList Scene::populatePartList() const
     renderer::populatePartList(partList, mRoot, mRoot.mInstance.mPose, 
                                mRoot.mInstance.mMaterialOverride ? &*mRoot.mInstance.mMaterialOverride : nullptr);
     return partList;
-}
-
-
-HardcodedUbos::HardcodedUbos(Storage & aStorage)
-{
-    aStorage.mUbos.emplace_back();
-    mFrameUbo = &aStorage.mUbos.back();
-    mUboRepository.emplace(semantic::gFrame, mFrameUbo);
-
-    aStorage.mUbos.emplace_back();
-    mViewingUbo = &aStorage.mUbos.back();
-    mUboRepository.emplace(semantic::gView, mViewingUbo);
-
-    aStorage.mUbos.emplace_back();
-    mModelTransformUbo = &aStorage.mUbos.back();
-    mUboRepository.emplace(semantic::gLocalToWorld, mModelTransformUbo);
 }
 
 
@@ -707,24 +333,23 @@ void registerGlfwCallbacks(graphics::AppInterface & aAppInterface,
 }
 
 
-RenderGraph::RenderGraph(const std::shared_ptr<graphics::AppInterface> aGlfwAppInterface,
+RenderGraph::RenderGraph(std::shared_ptr<graphics::AppInterface> aGlfwAppInterface,
                          const std::filesystem::path & aSceneFile,
                          const imguiui::ImguiUi & aImguiUi) :
     mGlfwAppInterface{std::move(aGlfwAppInterface)},
     // TODO How do we handle the dynamic nature of the number of instance that might be renderered?
     // At the moment, hardcode a maximum number
-    mInstanceStream{makeInstanceStream(mStorage, gMaxDrawInstances)},
     mLoader{makeResourceFinder()},
-    mUbos{mStorage},
     mCameraControl{mGlfwAppInterface->getWindowSize(),
                    gInitialVFov,
                    Orbital{2/*initial radius*/}
-    }
+    },
+    mGraph{mGlfwAppInterface, mStorage}
 {
     registerGlfwCallbacks(*mGlfwAppInterface, mCameraControl, EscKeyBehaviour::Close, &aImguiUi);
     registerGlfwCallbacks(*mGlfwAppInterface, mFirstPersonControl, EscKeyBehaviour::Close, &aImguiUi);
 
-    mScene = mLoader.loadScene(aSceneFile, "effects/Mesh.sefx", mInstanceStream, mStorage);
+    mScene = mLoader.loadScene(aSceneFile, "effects/Mesh.sefx", mGraph.mInstanceStream, mStorage);
     /*const*/Node & model = mScene.mRoot.mChildren.front();
 
     // TODO Ad 2023/10/03: Sort out this bit of logic: remove hardcoded sections,
@@ -762,7 +387,7 @@ RenderGraph::RenderGraph(const std::shared_ptr<graphics::AppInterface> aGlfwAppI
 
     // Add basic shapes to the the scene
     Handle<Effect> simpleEffect = makeSimpleEffect(mStorage);
-    auto [triangle, cube] = loadTriangleAndCube(mStorage, simpleEffect, mInstanceStream);
+    auto [triangle, cube] = loadTriangleAndCube(mStorage, simpleEffect, mGraph.mInstanceStream);
     mScene.addToRoot(triangle);
     mScene.addToRoot(cube);
 }
@@ -786,33 +411,6 @@ void RenderGraph::update(float aDeltaTime)
 }
 
 
-// TODO Ad 2023/09/26: Could be splitted between the part list load (which should be valid accross passes)
-// and the pass cache load (which might only be valid for a single pass)
-// TODO Ad 2023/11/26: Even more, since the model transform is to world space (i.e. constant accross viewpoints)
-// and it is fetched from a buffer in the shaders, it could actually be pushed once per frame.
-// (and the pass would only index the transform for the objects it actually draws).
-void RenderGraph::loadDrawBuffers(const PartList & aPartList,
-                                  const PassCache & aPassCache)
-{
-    PROFILER_SCOPE_RECURRING_SECTION("load_draw_buffers", CpuTime, BufferMemoryWritten);
-
-    assert(aPassCache.mDrawInstances.size() <= gMaxDrawInstances);
-
-    proto::load(*mUbos.mModelTransformUbo,
-                std::span{aPartList.mInstanceTransforms},
-                graphics::BufferHint::DynamicDraw);
-
-    // TODO Ad 2023/11/23: this hardcoded access to first buffer view is ugly
-    proto::load(*mInstanceStream.mVertexBufferViews.at(0).mGLBuffer,
-                std::span{aPassCache.mDrawInstances},
-                graphics::BufferHint::DynamicDraw);
-
-    proto::load(mIndirectBuffer,
-                std::span{aPassCache.mDrawCommands},
-                graphics::BufferHint::DynamicDraw);
-}
-
-
 void RenderGraph::drawUi()
 {
     mSceneGui.present(mScene);
@@ -825,6 +423,9 @@ void RenderGraph::render()
     nvtx3::mark("RenderGraph::render()");
     NVTX3_FUNC_RANGE();
 
+    auto & mUbos = mGraph.mUbos;
+    auto & mIndirectBuffer = mGraph.mIndirectBuffer;
+
     // Empty placeholder for the moment.
     RepositoryTexture dummyTextureRepository;
 
@@ -836,6 +437,8 @@ void RenderGraph::render()
     {
         PROFILER_SCOPE_RECURRING_SECTION("load_frame_UBOs", CpuTime, GpuTime, BufferMemoryWritten);
         loadFrameUbo(*mUbos.mFrameUbo);
+        // Note in a more realistic application, several cameras would be used per frame.
+        loadCameraUbo(*mUbos.mViewingUbo, mCamera);
     }
 
     // Use the same indirect buffer for all drawings
@@ -843,114 +446,7 @@ void RenderGraph::render()
     // With the current approach, this binding could be done only once at construction
     graphics::bind(mIndirectBuffer, graphics::BufferType::DrawIndirect);
 
-    // Hack until we have proper lifetime: give access to the texture outside the scope
-    const graphics::Texture * depthTexture;
-
-    // ----
-    // PASS (depth)
-    // ----
-    {
-        PROFILER_SCOPE_RECURRING_SECTION("pass_depth", CpuTime, GpuTime);
-
-        // Can be done once for distinct camera, if there is no culling
-        PassCache passCache = preparePass("depth", partList, mStorage);
-
-        // Load the data for the part and pass related UBOs (TODO: SSBOs)
-        loadDrawBuffers(partList, passCache);
-
-        //
-        // Attaching a depth texture to a FBO
-        //
-        const math::Size<2, int> depthMapSize = mGlfwAppInterface->getFramebufferSize();
-        // TODO should be a data member somewhere
-        static graphics::Texture depthMap = [](math::Size<2, int> aTextureSize, GLenum aFiltering)
-        {
-            graphics::Texture depthMap{GL_TEXTURE_2D};
-            graphics::allocateStorage(depthMap, GL_DEPTH_COMPONENT24, aTextureSize);
-            {
-                graphics::ScopedBind boundDepthMap{depthMap};
-                // TODO take sampling and filtering out, it is about reading the texture, not writing it.
-                glTexParameteri(depthMap.mTarget, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_BORDER);
-                glTexParameteri(depthMap.mTarget, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_BORDER);
-                glTexParameterfv(depthMap.mTarget, 
-                                 GL_TEXTURE_BORDER_COLOR,
-                                 math::hdr::Rgba_f{1.f, 0.f, 0.f, 0.f}.data());
-
-                glTexParameteri(depthMap.mTarget, GL_TEXTURE_MIN_FILTER, aFiltering);
-                glTexParameteri(depthMap.mTarget, GL_TEXTURE_MAG_FILTER, aFiltering);
-            }
-            return depthMap;
-        }(depthMapSize, GL_LINEAR/* seems required to get hardware pcf with sampler2DShadow */);
-
-        depthTexture = &depthMap;
-
-
-        //
-        // Set pipeline state
-        //
-        // Note: Must be set before any drawing operations, including glClear(),
-        // otherwise it becomes real mysterious real quick.
-        {
-            PROFILER_SCOPE_RECURRING_SECTION("set_pipeline_state", CpuTime);
-            glEnable(GL_CULL_FACE);
-            glEnable(GL_DEPTH_TEST);
-            glDepthMask(GL_TRUE);
-            glDisable(GL_BLEND);
-
-            // Viewport is coupled to the depth map here
-            glViewport(0, 0, depthMapSize.width(), depthMapSize.height());
-        }
-
-        static graphics::FrameBuffer depthFbo = [](const graphics::Texture & aTexture)
-        {
-            graphics::FrameBuffer depthFbo;
-            graphics::ScopedBind boundFbo{depthFbo};
-            // The attachment is permanent, no need to recreate it each time the FBO is bound
-            glFramebufferTexture2D(GL_DRAW_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, aTexture.mTarget, aTexture, /*mip map level*/0);
-            return depthFbo;
-        }(depthMap);
-        
-        graphics::ScopedBind boundFbo{depthFbo};
-
-        // Check FBO status before rendering
-        assert(glCheckFramebufferStatus(GL_DRAW_FRAMEBUFFER) == GL_FRAMEBUFFER_COMPLETE);
-
-        // Clear must appear after the Framebuffer setup!
-        gl.Clear(GL_DEPTH_BUFFER_BIT);
-
-        ///// TODO remove
-        static graphics::Texture colorTex = [](math::Size<2, int> aTextureSize)
-        {
-            graphics::Texture tex{GL_TEXTURE_2D};
-            graphics::allocateStorage(tex, GL_RGBA8, aTextureSize);
-            return tex;
-        }(depthMapSize);
-
-        glFramebufferTexture2D(GL_DRAW_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, colorTex.mTarget, colorTex, /*mip map level*/0);
-        const GLenum buffers[] = {GL_COLOR_ATTACHMENT0};
-        glDrawBuffers(1, buffers);
-
-        assert(glCheckFramebufferStatus(GL_DRAW_FRAMEBUFFER) == GL_FRAMEBUFFER_COMPLETE);
-
-        gl.Clear(GL_COLOR_BUFFER_BIT);
-        /////
-
-        // Might loop over cameras, or any other variation
-        //for(whatever dimension)
-        {
-            {
-                PROFILER_SCOPE_RECURRING_SECTION("load_pass_UBOs", CpuTime, GpuTime, BufferMemoryWritten);
-                // Note in a more realistic application, several cameras would be used per frame.
-                // TODO do not write it inside the pass if it can be done once above for several passes
-                loadCameraUbo(*mUbos.mViewingUbo, mCamera);
-            }
-
-            {
-                PROFILER_SCOPE_RECURRING_SECTION("draw_instances", CpuTime, GpuTime, GpuPrimitiveGen, DrawCalls, BufferMemoryWritten);
-                draw(passCache, mUbos.mUboRepository, dummyTextureRepository);
-            }
-        }
-    }
+    mGraph.passDepth(partList, mStorage);
 
     // ----
     // PASS (forward)
@@ -961,7 +457,7 @@ void RenderGraph::render()
         PassCache passCache = preparePass("forward", partList, mStorage);
 
         // Load the data for the part and pass related UBOs (TODO: SSBOs)
-        loadDrawBuffers(partList, passCache);
+        mGraph.loadDrawBuffers(partList, passCache);
 
         // FBO (default framebuffer)
         glBindFramebuffer(GL_DRAW_FRAMEBUFFER, 0);
@@ -985,12 +481,6 @@ void RenderGraph::render()
         // Might loop over cameras, or any other variation
         //for(whatever dimension)
         {
-            {
-                PROFILER_SCOPE_RECURRING_SECTION("load_pass_UBOs", CpuTime, GpuTime, BufferMemoryWritten);
-                // Note in a more realistic application, several cameras would be used per frame.
-                loadCameraUbo(*mUbos.mViewingUbo, mCamera);
-            }
-
             // TODO should be done once per viewport
             glViewport(0, 0,
                        mGlfwAppInterface->getFramebufferSize().width(),
@@ -1017,7 +507,7 @@ void RenderGraph::render()
 
         glBindFramebuffer(GL_DRAW_FRAMEBUFFER, 0);
         glActiveTexture(GL_TEXTURE0);
-        glBindTexture(depthTexture->mTarget, *depthTexture);
+        glBindTexture(mGraph.mDepthMap.mTarget, mGraph.mDepthMap);
 
         // Would require scissor test to clear only part of the screen.
         //gl.Clear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
@@ -1039,17 +529,6 @@ void RenderGraph::render()
         drawQuad({
             .mSourceChannel = 0,
             .mLinearization = DrawQuadParameters::DepthMethod1,
-            .mNearDistance = nearZ,
-            .mFarDistance = farZ,
-        });
-        
-        glViewport(0, fbSize.height() / 4,
-                   fbSize.width() / 4,
-                   fbSize.height() / 4);
-
-        drawQuad({
-            .mSourceChannel = 0,
-            .mLinearization = DrawQuadParameters::DepthMethod2,
             .mNearDistance = nearZ,
             .mFarDistance = farZ,
         });
