+++ src/apps/refactor_proto/refactor_proto/RendererReimplement.h
@@ -3,6 +3,8 @@
 
 #include "GlApi.h"
 
+#include <span>
+
 
 namespace ad::renderer {
